Complete rest of the logic.
Implemented a an until loop to be able to play the
game until either a win or tie happens.
Added a win_game? method to check if the board is
full or if there is a winner, and changed the player
class to have an array where the choices of players
are pushed into.